working on making battle engine - godmode
press Q when in the overworld to open the godmode menu. When in the godmode menu, you will eventually be able to change everything about the battles.
Currently, the only thing you can customize is the background color, but at least it is a start.
Q to open godmode, wasd and number keys to change godmode settings