room transitions
im working on room transitions right now. currently there is only the hall and one other room
also, when you end a battle, it will loop because the player's position is the same as it was when the battle started, which is touching the enemy, so the battle triggers again. i will fix this in the next version.
also lmk if the transition time should be shortened, it is very easy to do. it is currently 1 second