edit player position in godmode menu
i added the ability to change to position of the player in a room using the godmode menu. this can be used for level testing purposes, but the intention was that when you initiate a battle, each enemy will have a position for the player to be in after the battle set with it, and this will fix the issue of the looping battle, but it can't be used that way currently. i didnt explain that well, if you want me to explain it better, ask me on discord.
i also think i might have made some other changes but i forgot