fix sprite hit box size
i did this by creating a "top" layer in the rooms, and a top sprite for each npc.
when creating levels, make sure you have the main part of the npcs on the "instances" layer, and the top half overlapping it on the "assets top" layer for now.
soon, when i make npc dialogue, the top halves will go on the instances_top layer, but they go on assets_top for now.