can talk to an npc
working on npc dialogue.
currently the only person you can talk to is the emo kid in the top left corner of room 1
go up to them and press e to talk
press e again to quit talking
there is currently no visual cue for when you are close enough to talk to them, but i will do that later
also, the dialogue engine does not yet support multiple lines of dialogue, ill do that soon too.