Start of saving and loading
Press i when in the overworld and in a state where you are able to walk around to save
Press o under the same circumstances to load. Only a few pieces of data currently get carried over because I am still experimenting with saving and loading. it is still very buggy probably